+++ lizard_measure/api/urls.py
@@ -5,12 +5,13 @@ from django.contrib import admin
 
 from lizard_measure.api.views import RootView
 from lizard_measure.api.views import MeasureView
+from lizard_measure.api.views import WaterBodyView
 from lizard_measure.api.views import OrganizationView
 from lizard_measure.api.views import ScoreView
 from lizard_measure.api.views import SteeringParameterFreeView
 from lizard_measure.api.views import SteeringParameterPredefinedGraphView
 from lizard_measure.api.views import SteerParameterGraphs
-from lizard_measure.api.views import EsfPattern
+from lizard_measure.api.views import EsfPatternView
 from lizard_measure.api.views import SteerParameterOverview
 
 
@@ -26,6 +27,9 @@ urlpatterns = patterns(
     url(r'^measure/$',
         MeasureView.as_view(),
         name=NAME_PREFIX + 'measure'),
+    url(r'^waterbody/$',
+        WaterBodyView.as_view(),
+        name=NAME_PREFIX + 'water_body'),
     url(r'^organization/$',
         OrganizationView.as_view(),
         name=NAME_PREFIX + 'organization'),
@@ -41,13 +45,10 @@ urlpatterns = patterns(
     url(r'^steer_parameter_overview/$',
         SteerParameterOverview.as_view(),
         name=NAME_PREFIX + 'steer_parameter_overview'),
-
-
     url(r'^steer_parameter_graphs/$',
         SteerParameterGraphs.as_view(),
         name=NAME_PREFIX + 'steering_parameter_graphs'),
     url(r'^esf_pattern/$',
-        EsfPattern.as_view(),
+        EsfPatternView.as_view(),
         name=NAME_PREFIX + 'esf_pattern'),
-
     )