+++ lizard_measure/views.py
@@ -659,9 +659,10 @@ def steerparameter_overview(request):
     #predefined_graphs = ['testa', 'testb']
 
     parameters = SteeringParameterFree.objects.filter(area__in=areas).distinct().values_list('parameter_code', flat=True)
+
+    parameters = [{'org': par, 'no_point': par.replace('.','_')} for par in parameters]
     #parameters = ['test','test2']
-    if (request.user.has_perm('auth.is_analyst', SteeringParameterFree) and
-        request.user.has_perm('auth.is_analyst', PredefinedGraphSelection)):
+    if request.user.is_authenticated():
 
         t = get_template('portals/stuurparameter-overzicht.js')
         c = RequestContext(request, {